fix: error on redirecting output on unicode file paths (#211)
## How does this PR impact the user?

Before: When redirecting stdout to a file with unicode filepaths, the
program would raise an execption and crash

<img width="1098" height="289" alt="image"
src="https://github.com/user-attachments/assets/75708abf-d7c2-419b-8c8f-faa86fa8c45c"
/>

After: The program successfully writes output to file and displays
characters correctly

<img width="1084" height="200" alt="image"
src="https://github.com/user-attachments/assets/b76853fe-36f0-4587-aab2-3888a23460d1"
/>


## Description

Issue: #95 

Before outputting the scores and file paths of detected images, the
program will first check if it is outputting to console. If it is not
(therefore it is outputting to a file) it will change the encoding
method of stdout to `utf-8-sig` which encodes unicode correctly and
writes to the specified file using this encoding.

## Limitations

Have not tested impact on MacOS and Linux, however this should be none
to low.

## Checklist

- [x] my PR is focused and contains one holistic change
- [x] I have added screenshots or screen recordings to show the changes

P.S if you are happy with this PR, please mark it as
`hacktoberfest-accepted` :)